Fix PyExt error handling (#7042)
The current PythonExtensionGen code attempts to provide verbose error (exception) messages by overriding halide_error and saving the message in a thread_local. This isn't safe or correct, however, and (in general) is wrong for any Halide code using multiple threads. #6994 proposes ways to mitigate this (and there are experiments in place to implement it), but unless/until those enhancements land, we can't leave the code in its current state. So:
- Don't try to save the text at all.
- Optionally log the text to stderr.
- Just throw an exception with the numeric error code. This is suboptimal, but better than the existing usually-incorrect-message behavior.
- Bonus: wrap both the error and print overloads with `PyGILState_Ensure()`, as we are supposed to, to ensure we don't die.