Fix false positive use after free warning.
Latest clang is giving a use after free warning for printing a pointer
value after it is freed. The compiler is not exactly wrong in that a
pointer is being passed to something after it is freed, but the
something is just printing the pointer's value for debugging
purposes. Cast through an intptr_t to avoid the warning.