Attempt to fix Python layout once and for all
TL;DR: our current Python structure requires use of relative paths (not ideal) and a lot of fiddling with PYTHONPATH (extremely not ideal). This restructures things so that PYTHONPATH ~never has to be set, and all imports are via absolute Python package/module specifications, rather than relative. Most of this is just code motion.