Add "argv_signature" to C header generation
Add a new call, `$funcname_argv_signature()`, to our generated C header files. This returns a const char * one character per expected argument to the 'argv' call;
each character is used to indicate the type of the argument. The character encoding is (mostly) borrowed from Python/NumPy, with some minor extensions.

This is intended to provide a lightweight mechanism to write reusable marshal/unmarshal code across language/processor/etc boundaries. (Note that while the existing
metadata provides equivalent information, it is also heavier-weight to process, and hard to replicate from scratch; this, on the other hand, provides a convention
that can be constructed simply by inspection, or even copy/pasted into source files that may not easily be able to process halide metadata.)

Note that this is being implemented as an inline function, rather than a constant or a #define; I think this is the safest way to implement this in a way that works with C (C99 or later) as well as C++.

This doesn't need to land yet, but I'm about to try building something on top of it and if there are loud objections I'd rather hear sooner than later.

```
!   bfloat16
e   float16
f   float32
d   float64
b   int8
h   int16
i   int32
q   int64
?   bool
B   uint8
H   uint16
I   uint32
Q   uint64
P   void * / handle
@   halide_buffer_t * (as input)
#   halide_buffer_t * (as output)

```