Fix readnone attribute for llvm 16
The readnone flag was changed to memory(none) when applied to functions.
llvm-as dynamically upgrades readnone applied to functions, so our .ll
is fine for now, but there were places in the compiler we were manually
sticking 'readnone' on a function.

Also did a driveby makefile fix to remove some vestigial wasm stuff that
was throwing errors with newer versions of llvm-config