Revise simd_op_check tests to ignore HL_TARGET (#7207) (#7216)
* Revise simd_op_check tests to ignore HL_TARGET (#7207)

The simd_op_check tests have historically only run using the value of HL_TARGET, which mean that the coverage they had was low (since HL_TARGET is only set to values that are runnable on at least one buildbot). This change completely disconnects these tests from HL_TARGET; instead, each test now tests for a range of targets appropriate to the architecture being tested. On all platforms, they still compile to assembly and verify that the correct instructions are generated; additionally, if the host platform can JIT for the given target, it verifies that the results are as expected.

* Update simd_op_check_riscv.cpp

* Update simd_op_check_x86.cpp

* Update simd_op_check_x86.cpp

* Update simd_op_check_arm.cpp

* Add more features that must match; re-enable the bfloat instructions

* Update simd_op_check_x86.cpp

* Update simd_op_check_riscv.cpp

* trigger buildbots

* Fix simd_op_check_wasm