Fix declaration of extern fns in C backend that use __user_context
TL;DR: the headers declare it `void const *` but our internal declarations declared it `void *`; if we had a C++-backend piece of code that had a define_extern to another C++-backend piece of code (and both were C++ and both used user_context), the name mangling was wrong and linking would fail. Whew.