Unify user_context name-and-type everywhere
For reasons that aren't clear, we sometimes emitted user_context as `void*` or `Handle()` (most places), but sometimes as `const void *` (Codegen_C). This attempts to unify it to be plain `void *` everywhere, to avoid the need for const-cast and also void possibilities of c++ name-mangling linker mismatches in some obscure cases.

This is technically a breaking change, since existing code may be passing in `const void *` pointers; I think this is arguably acceptable, since the breakage will be very obvious and very straightforward to fix.

Alternately, I considered unifying on `const void *`, but that seems more restrictive at the call site, and makes for a weird mismatch with our runtime functions (which are all uniformly `void *`).

I'm not entirely happy with the breakage, but the status quo is clearly wrong, and really, has always been.