(V2) Unify user_context name-and-type everywhere
For reasons that aren't clear, we sometimes emitted user_context as `void*` or `Handle()` (most places), but sometimes as `const void *` (Codegen_C). This attempts to unify it to be `const void *` for all user-generated code, to avoid the need for const-cast and also void possibilities of c++ name-mangling linker mismatches in some obscure cases.