Add CMake support for C++ backend in test/generator (#7274)
* Add support for C++ backend in test/generator

When the CMake rules were rewritten a while back, the support for building/testing generators with the C++ backend (instead of the standard LLVM, etc) got lost. This adds it back in.

Also made some drive-by fixes to the Makefile to enable some tests there that work correctly now.

Also made a drive-by fix in in Codegen_C to fix allocation nodes that were just wrappers around buffer_get_host -- this prevented the cleanup_on_error test from building with the C++ backend.