Remove ParamMap (Fix #7121)
This PR proposes to remove ParamMap entirely for Halide 16; it was added to provide a threadsafe way to provide parameteres to the JIT, but `compile_to_callable()` now does this in a much less intrusive way.

Normally I'd propose deprecating something like this first (eg via ifdef), but in this case it would be extremely intrusive to do so, so this PR just rips off the band-aid and removes it entirely -- feedback welcome on this.