Attempt to give proper labels to pip 'nighty' builds
Uploading to the test server requires unique names. This attempts to use versions of the form `devTIME`, where TIME is unix-time-UTC. This PR is arguably incomplete right now (I suspect it would be better to switch to 'true' nightlies that trigger at, say, 9pm PST, and also to add rotation so we only keep a handful of uploads); this is meant to test the basic mechanism to see if the GitHub Action setup will work.