Always use __atomic in runtime thread operations
Upstream LLVM is strongly suggesting we move away from __sync primitives and use __atomic primitives instead.

IIRC, we still use the older __sync primitives on 32-bit because of some tooling issues -- perhaps on arm32 linux?

This PR is really just a test to see if those issues are still the case on our current set of buildbots.

(There are other `__sync` primitives usages in Halide that would need to be migrated as well if we decide to migrate, this is just an initial test)