Use existing Halide Runtime atomic wrappers everywhere in the runtime
We currently have a set of wrappers around the __atomic/__sync primitives used by our threading model; for various reasons, we desire to use the (deprecated) __sync primitives for 32-bit builds instead of the __atomic primitives (see https://github.com/halide/Halide/pull/7427 for some discussion).

This PR attempts to use these abstractions everywhere else in our runtime, for consistency, so that 64-bit builds consistent use the __atomic primitives for (e.g.) profiling and tracing too. This meant:
- Splitting the wrappers into a new header (runtime_atomics.h), as synchronization_common.h can't be included into arbitrary other files, for valid reasons
- Adding wrappers for the necessary primitives
- Modifying the code elsewhere in the runtime

Where new wrappers were needed, I generally defaulted to assuming that SEQ_CST was the safest memory order to use.

Not entirely sure if this is a worthwhile goal or not, but putting this out there for consideration and discussion.