Rename `device_copy` -> `DeviceCopy`
This is 100% stylistic, but it's bothered me for years: the `device_copy` struct doesn't follow the usual Halide style for struct types, so every time I see it, I want to read it as a function name. Let's rename it to DeviceCopy to help keep me sane?

Also, drive-by change of #define to constexpr since we have C++17 in runtime.