Completely rework how RISC-V vector intrinsics are called to avoid issues iwth single element vectors being confused with scalars and other conversions that can happen via using call_intrin.
Allows using any size vector. Only downside is splitting large vectors
no longer happens, but RISC-V allows an LMUL of 8, meaning a vector of
up to 8 times the vector register size will compile so this is much
less of an issue. Splitting larger vectors can be added.

Should also allow fractionaly LMUL in all cases, but this is not
verified.