Backport #7650 to the release/16.x branch (#7653)
Default RISCV backend to OFF for LLVM < 17 (#7650)

LLVM17 is doing a lot of work on the RISCV backend, and the amount of testing done on Halide's LLVM16-based RISCV codegen is very light. It's been suggested that we should default to not enabling the RISCV backend for LLVM16 and earlier because of this (so that people attempting to use Halide for RISCV won't encounter a possible footgun).

This PR just adds the relevant mechanism; whether or not this is the correct decision is not clear. Discussion welcome.