Fix correctness_float16_t for ASAN builds
This appears to be a glitch that has to do with changing ABI for float16 across versions of GCC; we build LLVM with gcc-9 on Linux, but the float16 ABI got changed (and unified in gcc12); since ASAN builds use Clang even on linux, there is a hiccup here.

This is an ugly monkey-patch to work around this issue.