Revise labels on autoscheduler tests
This is step 1 in fixing https://github.com/halide/Halide/issues/7731: it replaces the `autoschedulers` tag with more granular ones, so that we can modify the build script to test the right autoscheduler(s) for a given backend. (Note that the `autoschedulers` tag was unused by the buildbots, which only used the generic `auto_schedule` tag.)

Step 2 will be to modify the buildbot script after this lands to use the new tags above.

Step 3 will be to remove the `auto_schedule` tag.