Unbreak WebGPU builds
It seems that newer versions of emcc and/or dawn no longer have the `wgpuDeviceSetDeviceLostCallback()` fn available. This just comments these out temporarily, to unbreak the builds; it's not clear if these are no longer necessary or if there is an alternate API to accomplish this, but while I'm investigating, this is better than nothing.