Make sure all Halide arithmetic scalar types can be named from the Generator interface. (#7934)
* Make sure all Halide arithmetic scalar types can be named from the
Generator interface. Specifically adding 64-bit signed and unsigned
integers and making sure float16 and bfloat16 are fully supported and
documented.

Add a simple test for all the type names.

(Don't use float16 and bfloat16 in the arithmetic as they do not compile
with the C++ backend. The name mapping should still be tested but the
types passed do not seem to be checked as the values are not used.)